from float to double

+++ source/Initializers.h
@@ -3,7 +3,7 @@
 
 using namespace std;
 
-void initializeRandomMatrix(float***, int, mt19937*);
-void initializeNullMatrix(float***, int);
-void initializeRandomVector(float**, int, mt19937*);
-void initializeNullVector(float**, int);
+void initializeRandomMatrix(double***, int, mt19937*);
+void initializeNullMatrix(double***, int);
+void initializeRandomVector(double**, int, mt19937*);
+void initializeNullVector(double**, int);
